Fix xs:element fixed value cascade

+++ output/models/ibm_data/valid/s3_12/s3_12v04_xsd/s3_12v04.py
@@ -59,7 +59,7 @@ class Root:
         name = "root"
         namespace = "http://xstest-tns"
 
-    elem1: List[Union[AnySimpleType, DoubleType, FloatType]] = field(
+    elem1: List[Union[AnySimpleType, FloatType, DoubleType]] = field(
         default_factory=list,
         metadata={
             "type": "Element",
